+++ spec/integrations/factory_doctor_spec.rb
@@ -15,7 +15,7 @@ describe "FactoryDoctor" do
       expect(output).to include("generates random names (./factory_doctor_fixture.rb:13) – 2 records created")
       expect(output).to include("validates name (./factory_doctor_fixture.rb:18) – 1 record created")
       expect(output).to include("clones (./factory_doctor_fixture.rb:28) – 1 record created")
-      expect(output).to include("PlainMinitest (./factory_doctor_fixture.rb)")
+      expect(output).to include("PlainMinitestFabricationTest (./factory_doctor_fixture.rb)")
       expect(output).not_to include("is ignored")
       expect(output).not_to include("creates and reloads user")
     end
@@ -33,7 +33,7 @@ describe "FactoryDoctor" do
       output = run_rspec("factory_doctor", env: {"FDOC" => "1"})
 
       expect(output).to include("FactoryDoctor report")
-      expect(output).to include("Total (potentially) bad examples: 3")
+      expect(output).to include("Total (potentially) bad examples: 4")
       expect(output).to match(/Total wasted time: \d{2}:\d{2}\.\d{3}/)
 
       expect(output).to include("User (./factory_doctor_fixture.rb:7)")
@@ -71,13 +71,13 @@ describe "FactoryDoctor" do
           env: {"FDOC" => "1", "FDOC_STAMP" => "fd_ignore"}
         )
 
-        expect(output).to include("5 examples, 0 failures")
+        expect(output).to include("7 examples, 0 failures")
 
         expect(output).to include("FactoryDoctor report")
-        expect(output).to include("Total (potentially) bad examples: 3")
+        expect(output).to include("Total (potentially) bad examples: 4")
 
         expect(output).to include("RSpec Stamp results")
-        expect(output).to include("Total patches: 3")
+        expect(output).to include("Total patches: 4")
         expect(output).to include("Total files: 1")
         expect(output).to include("Failed patches: 0")
         expect(output).to include("Ignored files: 0")
@@ -87,7 +87,7 @@ describe "FactoryDoctor" do
           env: {"FDOC" => "1"}
         )
 
-        expect(output2).to include("5 examples, 0 failures")
+        expect(output2).to include("7 examples, 0 failures")
         expect(output2).to include('FactoryDoctor says: "Looks good to me!"')
       end
     end